moved WindowsType enum class in Events pkg
